+++ src/main/java/com/imsweb/naaccrxml/sas/SasCsvToXml.java
@@ -252,7 +252,7 @@ public class SasCsvToXml {
                         if (_specificationsVersion != null)
                             writer.write("\n            specificationVersion=\"" + _specificationsVersion + "\"");
                         else
-                            writer.write("\n            specificationVersion=\"1.7\"");
+                            writer.write("\n            specificationVersion=\"1.8\"");
                         writer.write("\n            xmlns=\"http://naaccr.org/naaccrxml\"");
                         writer.write(">\n");
                         for (Entry<String, String> entry : rootFields.entrySet()) {